corrections based on Alessandro's suggestions
