+++ src/main/java/org/isf/opd/service/OpdIoOperationRepositoryImpl.java
@@ -36,40 +36,24 @@ import javax.persistence.criteria.Root;
 
 import org.isf.generaldata.MessageBundle;
 import org.isf.opd.model.Opd;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
 public class OpdIoOperationRepositoryImpl implements OpdIoOperationRepositoryCustom {
-	
+
 	@PersistenceContext
 	private EntityManager entityManager;
 
-	@SuppressWarnings("unchecked")	
+	@SuppressWarnings("unchecked")
 	@Override
-	public List<Opd> findAllOpdWhereParams(
-			String diseaseTypeCode,
-			String diseaseCode, 
-			GregorianCalendar dateFrom,
-			GregorianCalendar dateTo,
-			int ageFrom, 
-			int ageTo,
-			char sex,
-			char newPatient) {
-		return _getOpdQuery(
-						diseaseTypeCode, diseaseCode, dateFrom, dateTo,
-						ageFrom, ageTo, sex, newPatient).
-					getResultList();
-	}	
+	public List<Opd> findAllOpdWhereParams(String diseaseTypeCode, String diseaseCode, GregorianCalendar dateFrom, GregorianCalendar dateTo, int ageFrom,
+					int ageTo, char sex, char newPatient) {
+		return _getOpdQuery(diseaseTypeCode, diseaseCode, dateFrom, dateTo, ageFrom, ageTo, sex, newPatient).getResultList();
+	}
 
-	private TypedQuery<Opd> _getOpdQuery(
-			String diseaseTypeCode,
-			String diseaseCode, 
-			GregorianCalendar dateFrom,
-			GregorianCalendar dateTo,
-			int ageFrom, 
-			int ageTo,
-			char sex,
-			char newPatient) {
+	private TypedQuery<Opd> _getOpdQuery(String diseaseTypeCode, String diseaseCode, GregorianCalendar dateFrom, GregorianCalendar dateTo, int ageFrom,
+					int ageTo, char sex, char newPatient) {
 		CriteriaBuilder cb = entityManager.getCriteriaBuilder();
 		CriteriaQuery<Opd> query = cb.createQuery(Opd.class);
 		Root<Opd> opd = query.from(Opd.class);
@@ -77,35 +61,25 @@ public class OpdIoOperationRepositoryImpl implements OpdIoOperationRepositoryCus
 
 		query.select(opd);
 		if (!(diseaseTypeCode.equals(MessageBundle.getMessage("angal.common.alltypes.txt")))) {
-			predicates.add(
-				cb.equal(opd.join("disease").join("diseaseType").get("code"), diseaseTypeCode)
-			);
+			predicates.add(cb.equal(opd.join("disease").join("diseaseType").get("code"), diseaseTypeCode));
 		}
 		if (!diseaseCode.equals(MessageBundle.getMessage("angal.opd.alldiseases.txt"))) {
-			predicates.add(
-				cb.equal(opd.join("disease").get("code"), diseaseCode)
-			);
+			predicates.add(cb.equal(opd.join("disease").get("code"), diseaseCode));
 		}
 		if (ageFrom != 0 || ageTo != 0) {
-			predicates.add(
-				cb.between(opd.<Integer>get("age"), ageFrom, ageTo)
-			);
+			predicates.add(cb.between(opd.<Integer> get("age"), ageFrom, ageTo));
 		}
 		if (sex != 'A') {
-			predicates.add(
-				cb.equal(opd.get("sex"), sex)
-			);
+			predicates.add(cb.equal(opd.get("sex"), sex));
 		}
 		if (newPatient != 'A') {
-			predicates.add(
-				cb.equal(opd.get("newPatient"), newPatient)
-			);
+			predicates.add(cb.equal(opd.get("newPatient"), newPatient));
 		}
-		predicates.add(
-			cb.between(opd.<Date>get("visitDate"), dateFrom.getTime(), dateTo.getTime())
-		);
+		predicates.add(cb.between(opd.<Date> get("visitDate"), dateFrom.getTime(), dateTo.getTime()));
 		query.where(cb.and(predicates.toArray(new Predicate[predicates.size()])));
 
 		return entityManager.createQuery(query);
 	}
+
+	
 }
