+++ src/main/java/org/isf/patient/service/PatientIoOperations.java
@@ -37,10 +37,20 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 /**
- * ------------------------------------------ PatientIoOperations - dB operations for the patient entity ----------------------------------------- modification
- * history 05/05/2005 - giacomo - first beta version 03/11/2006 - ross - added toString method. Gestione apici per nome, cognome, citta', indirizzo e note
- * 11/08/2008 - alessandro - added father & mother's names 26/08/2008 - claudio - added birth date modified age 01/01/2009 - Fabrizio - changed the calls to
- * PAT_AGE fields to return again an int type 03/12/2009 - Alex - added method for merge two patients history ------------------------------------------
+ * ------------------------------------------
+ * PatientIoOperations - dB operations for the patient entity
+ * -----------------------------------------
+ * modification history
+ * 05/05/2005 - giacomo  - first beta version
+ * 03/11/2006 - ross - added toString method. Gestione apici per
+ * nome, cognome, citta', indirizzo e note
+ * 11/08/2008 - alessandro - added father & mother's names
+ * 26/08/2008 - claudio    - added birth date
+ * modified age
+ * 01/01/2009 - Fabrizio   - changed the calls to PAT_AGE fields to
+ * return again an int type
+ * 03/12/2009 - Alex       - added method for merge two patients history
+ * ------------------------------------------
  */
 @Service
 @Transactional(rollbackFor = OHServiceException.class)
@@ -61,16 +71,6 @@ public class PatientIoOperations {
 	 */
 	public ArrayList<Patient> getPatients() throws OHServiceException {
 		return new ArrayList<>(repository.findByDeletedOrDeletedIsNull(NOT_DELETED_STATUS));
-	}
-
-	/**
-	 * Count all active patients
-	 * 
-	 * @return
-	 * @throws OHServiceException
-	 */
-	public long countAllActivePatients() throws OHServiceException {
-		return repository.countAllActivePatients();
 	}
 
 	/**
@@ -106,8 +106,7 @@ public class PatientIoOperations {
 	 * - taxCode<br>
 	 * - note<br>
 	 *
-	 * @param keyword
-	 *            - String to search, <code>null</code> for full list
+	 * @param keyword - String to search, <code>null</code> for full list
 	 * @return the list of Patients (could be empty)
 	 * @throws OHServiceException
 	 */
@@ -177,8 +176,7 @@ public class PatientIoOperations {
 	/**
 	 * Method that updates an existing {@link Patient} in the db
 	 *
-	 * @param patient
-	 *            - the {@link Patient} to update
+	 * @param patient - the {@link Patient} to update
 	 * @return true - if the existing {@link Patient} has been updated
 	 * @throws OHServiceException
 	 */
@@ -199,8 +197,8 @@ public class PatientIoOperations {
 	}
 
 	/**
-	 * Method that check if a Patient is already present in the DB by his/her name (the passed string 'name' should be a concatenation of firstName + " " +
-	 * secondName
+	 * Method that check if a Patient is already present in the DB by his/her name
+	 * (the passed string 'name' should be a concatenation of firstName + " " + secondName
 	 *
 	 * @param name
 	 * @return true - if the patient is already present
@@ -239,13 +237,23 @@ public class PatientIoOperations {
 	/**
 	 * Checks if the code is already in use
 	 *
-	 * @param code
-	 *            - the patient code
+	 * @param code - the patient code
 	 * @return <code>true</code> if the code is already in use, <code>false</code> otherwise
 	 * @throws OHServiceException
 	 */
 	public boolean isCodePresent(Integer code) throws OHServiceException {
 		return repository.exists(code);
+	}
+	
+	
+	/**
+	 * Count all active patients
+	 * 
+	 * @return
+	 * @throws OHServiceException
+	 */
+	public long countAllActivePatients() throws OHServiceException {
+		return repository.countAllActivePatients();
 	}
 
 }
