+++ src/main/java/org/isf/telemetry/envdatacollector/AbstractDataCollector.java
@@ -19,20 +19,22 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.isf.remote.iptools.geoip.freegeoip;
+package org.isf.telemetry.envdatacollector;
 
-import org.springframework.cloud.netflix.feign.FeignClient;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
+import java.util.Map;
 
-@FeignClient(name = "freegeoip-remote-service")
-public interface FreeGeoIPRemoteService {
+import org.isf.utils.exception.OHException;
 
-	// @formatter:off
+public abstract class AbstractDataCollector {
 
-	@GetMapping(value = "/json", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<FreeGeoIPJSON> retrieveGeoIPInfo();
+	public abstract String getId();
 
-	// @formatter:on
+	public abstract String getDescription();
+
+	public abstract Map<String, String> retrieveData() throws OHException;
+
+	public boolean isSelected(Map<String, Boolean> checkboxesStatus) {
+		return getId() != null && checkboxesStatus.containsKey(this.getId()) && Boolean.TRUE.equals(checkboxesStatus.get(this.getId()));
+	}
+
 }
