+++ src/main/java/org/isf/telemetry/envdatacollector/DataCollectorProviderService.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.isf.datacollector;
+package org.isf.telemetry.envdatacollector;
 
 import java.util.HashMap;
 import java.util.List;
