+++ src/main/java/org/isf/telemetry/envdatacollector/collectors/ApplicationDataCollector.java
@@ -19,31 +19,32 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.isf.datacollector.collectors;
+package org.isf.telemetry.envdatacollector.collectors;
 
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Properties;
 
-import org.isf.datacollector.AbstractDataCollector;
-import org.isf.datacollector.constants.CollectorsConst;
-import org.isf.remote.iptools.geoip.freegeoip.FreeGeoIPJSON;
-import org.isf.remote.iptools.geoip.freegeoip.FreeGeoIPService;
+import org.isf.generaldata.Version;
+import org.isf.telemetry.envdatacollector.AbstractDataCollector;
+import org.isf.telemetry.envdatacollector.constants.CollectorsConst;
+import org.isf.utils.db.DbSingleConn;
 import org.isf.utils.exception.OHException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
-@Order(value = 60)
+@Order(value = 10)
 @Component
-public class LocationDataCollector extends AbstractDataCollector {
+public class ApplicationDataCollector extends AbstractDataCollector {
 
-	private static final String ID = "FUN_LOCATION";
-	private static final Logger LOGGER = LoggerFactory.getLogger(LocationDataCollector.class);
-
-	@Autowired
-	private FreeGeoIPService freeGeoIpService;
+	private static final String ID = "FUN_APPLICATION";
+	private static final Logger LOGGER = LoggerFactory.getLogger(ApplicationDataCollector.class);
 
 	@Override
 	public String getId() {
@@ -52,28 +53,24 @@ public class LocationDataCollector extends AbstractDataCollector {
 
 	@Override
 	public String getDescription() {
-		return "Location information (ex. Italy)";
+		return "Application technical information (ex. OH version)";
 	}
 
 	@Override
 	public Map<String, String> retrieveData() throws OHException {
-		LOGGER.debug("Collecting location data...");
+		LOGGER.debug("Collecting application data...");
 		Map<String, String> result = new HashMap<>();
 		try {
-			FreeGeoIPJSON json = this.freeGeoIpService.retrieveGeoIpInfo();
-			result.put(CollectorsConst.LOC_COUNTRY_NAME, json.getCountryName());
-			result.put(CollectorsConst.LOC_COUNTRY_CODE, json.getCountryCode());
-			result.put(CollectorsConst.LOC_REGION_NAME, json.getRegionName());
-			result.put(CollectorsConst.LOC_REGION_CODE, json.getRegionCode());
-			result.put(CollectorsConst.LOC_CITY, json.getCity());
-			result.put(CollectorsConst.LOC_ZIP_CODE, json.getZipCode());
-			result.put(CollectorsConst.LOC_TIMEZONE, json.getTimeZone());
-		} catch (RuntimeException e) {
+			Version.initialize();
+			result.put(CollectorsConst.APP_VER_MAJOR, Version.VER_MAJOR);
+			result.put(CollectorsConst.APP_VER_MINOR, Version.VER_MINOR);
+			result.put(CollectorsConst.APP_RELEASE, Version.VER_RELEASE);
+			return result;
+		} catch (Exception e) {
 			LOGGER.error("Something went wrong with " + ID);
 			LOGGER.error(e.toString());
 			throw new OHException("Data collector [" + ID + "]", e);
 		}
-		return result;
 	}
 
 }
