+++ src/main/java/org/isf/telemetry/envdatacollector/constants/CollectorsConst.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.isf.datacollector.constants;
+package org.isf.telemetry.envdatacollector.constants;
 
 public interface CollectorsConst {
