--- src/main/java/org/isf/utils/time/DateUtils.java
@@ -1,38 +1,0 @@
-/*
- * Open Hospital (www.open-hospital.org)
- * Copyright © 2006-2021 Informatici Senza Frontiere (info@informaticisenzafrontiere.org)
- *
- * Open Hospital is a free and open source software for healthcare data management.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * https://www.gnu.org/licenses/gpl-3.0-standalone.html
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-package org.isf.utils.time;
-
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
-public class DateUtils {
-
-	private static final String DATE_FORMAT = "yyyy-MM-dd";
-
-	public static boolean isNotToday(Date date) {
-		SimpleDateFormat sdf = new SimpleDateFormat(DATE_FORMAT);
-		String yesterday = sdf.format(date);
-		String today = sdf.format(new Date());
-		return date == null || !today.equals(yesterday);
-	}
-
-}
